initial upward and downward tile flicking
