+++ frontend/src/app/demo3/page.tsx
@@ -2,25 +2,12 @@
 
 import { useEffect, useRef } from "react";
 import * as PIXI from "pixi.js";
+import { getLayoutConstants, tweenTo } from "./utils";
+import { createSlotHelpers } from "./slots";
+import { createTileHelpers } from "./tiles";
 
 export default function Page() {
   const containerRef = useRef<HTMLDivElement>(null);
-  function getLayoutConstants() {
-    const width = window.innerWidth;
-
-    // Smooth proportional scale for tiles/slots
-    const scale = Math.min(Math.max(width / 1200, 0.5), 1);
-
-    // Scale slot and gap proportionally
-    const SLOT_SIZE = 80 * scale;
-    const SLOT_GAP = 16 * scale;
-
-    // Padding should *not* shrink as much — interpolate between 50 and 32 instead
-    const PADDING = 50 - (width - 400) * (18 / 800); // transitions from 50 → 32
-    const clampedPadding = Math.max(32, Math.min(50, PADDING));
-
-    return { PADDING: clampedPadding, SLOT_SIZE, SLOT_GAP };
-  }
 
   useEffect(() => {
     let app: PIXI.Application | null = null;
@@ -49,9 +36,7 @@ export default function Page() {
       let { PADDING, SLOT_SIZE, SLOT_GAP } = getLayoutConstants();
       const LETTERS = ["A", "B", "C", "D", "E", "F", "G"]; // letters in rack
 
-      // --- Vertical positions
-      const SLOTS_Y = () => PADDING; // slots row near top
-      const RACK_Y = () => app!.renderer.height - PADDING - SLOT_SIZE - 20; // rack row near bottom
+    
 
       // --- Layers for better z-ordering
       const slotsLayer = new PIXI.Container();
@@ -67,287 +52,42 @@ export default function Page() {
         fontWeight: "800",
       });
 
-      // --- Helpers for animations
-      const lerp = (a: number, b: number, t: number) => a + (b - a) * t;
-      const clamp = (n: number, lo: number, hi: number) =>
-        Math.max(lo, Math.min(hi, n));
-
-      // Smooth tweening for tile movement
-      function tweenTo(
-        obj: any,
-        to: { x: number; y: number },
-        ms = 300,
-        onComplete?: () => void
-      ) {
-        const from = { x: obj.x, y: obj.y };
-        const start = performance.now();
-        function step(now: number) {
-          const t = clamp((now - start) / ms, 0, 1);
-          // easeInOutQuad curve
-          const ease = t < 0.5 ? 2 * t * t : -1 + (4 - 2 * t) * t;
-          obj.x = lerp(from.x, to.x, ease);
-          obj.y = lerp(from.y, to.y, ease);
-          if (t < 1) requestAnimationFrame(step);
-          else onComplete && onComplete();
-        }
-        requestAnimationFrame(step);
-      }
+      // animation helpers moved to ./utils.ts
 
       // =========================
-      // SLOTS SETUP
+      // SLOTS SETUP (moved to ./slots)
       // =========================
-      let slots: any[] = [];
-
-      function layoutSlots() {
-        slots.length = 0;
-        slotsLayer.removeChildren();
-
-        const totalWidth = 7 * SLOT_SIZE + 6 * SLOT_GAP;
-        const startX = (app!.renderer.width - totalWidth) / 2;
-
-        for (let i = 0; i < 7; i++) {
-          const g = new PIXI.Graphics();
-          g.roundRect(0, 0, SLOT_SIZE, SLOT_SIZE, 16)
-            .fill({color:0xe5e7eb}) // Tailwind gray-200
-            .stroke({ color: 0x94a3b8, width: 2 }); // Tailwind gray-400
-          g.alpha = 0.9;
-          g.x = Math.round(startX + i * (SLOT_SIZE + SLOT_GAP));
-          g.y = Math.round(SLOTS_Y());
-          g.eventMode = "none"; // slots are not interactive
-          slotsLayer.addChild(g);
-          slots.push({ g, x: g.x, y: g.y, occupiedBy: null, index: i });
-        }
-      }
-
-      // Rack positions for letters
-      function rackPositions() {
-        const totalWidth =
-          LETTERS.length * SLOT_SIZE + (LETTERS.length - 1) * SLOT_GAP;
-        const startX = (app!.renderer.width - totalWidth) / 2;
-        return LETTERS.map((_, i) => ({
-          x: Math.round(startX + i * (SLOT_SIZE + SLOT_GAP)),
-          y: Math.round(RACK_Y()),
-        }));
-      }
+      const {
+        slots,
+        layoutSlots,
+        rackPositions,
+        hitTestEmptySlot,
+        nearestEmptySlot,
+        firstAvailableSlot,
+      } = createSlotHelpers({
+        app: app!,
+        SLOT_SIZE,
+        SLOT_GAP,
+        PADDING,
+        slotsLayer,
+        LETTERS,
+      });
 
       // =========================
-      // TILES SETUP
+      // TILES SETUP (moved to ./tiles)
       // =========================
-      let tiles: any[] = [];
-
-      function createTile(letter: string, idx: number) {
-        const c = new PIXI.Container();
-        const g = new PIXI.Graphics();
-        g.roundRect(0, 0, SLOT_SIZE, SLOT_SIZE, 16)
-          .fill({color:0xfef3c7}) // Tailwind amber-100
-          .stroke({ color: 0xf59e0b, width: 3 }); // Tailwind amber-500
-        const t = new PIXI.Text({ text: letter, style: textStyle });
-        t.anchor.set(0.5);
-        t.x = SLOT_SIZE / 2;
-        t.y = SLOT_SIZE / 2;
-        c.addChild(g, t);
-        c.cursor = "grab";
-        c.eventMode = "static"; // enable pointer events
-        c.zIndex = 1;
-        tilesLayer.addChild(c);
-
-        const home = rackPositions()[idx]; // initial rack position
-        c.position.set(home.x, home.y);
-
-        const tile: any = {
-          ctr: c,
-          label: t,
-          letter,
-          home: { ...home },
-          slot: null, // slot reference if placed
-          vx: 0,
-          vy: 0,
-          lastMoves: [] as any[], // history for flick detection
-        };
-        enableTileInteraction(tile);
-        tiles.push(tile);
-        return tile;
-      }
+      const { tiles, createTile, placeTileInSlot, returnTileHome } = createTileHelpers({
+        SLOT_SIZE,
+        tilesLayer,
+        textStyle,
+        rackPositions,
+        hitTestEmptySlot,
+        nearestEmptySlot,
+        firstAvailableSlot,
+      });
 
       LETTERS.forEach((L, i) => createTile(L, i));
       tilesLayer.sortableChildren = true;
-
-      // =========================
-      // INTERACTION LOGIC
-      // =========================
-      function enableTileInteraction(tile: any) {
-        const c = tile.ctr;
-        let dragging = false;
-        let dragOffset = { x: 0, y: 0 };
-
-        // Click tracking
-        let lastClickTime = 0;
-        let clickTimeout: any = null;
-
-        // --- pointer down
-        c.on("pointerdown", (e: any) => {
-          dragging = true;
-          c.zIndex = 1000;
-          c.cursor = "grabbing";
-          const global = e.global;
-          dragOffset.x = global.x - c.x;
-          dragOffset.y = global.y - c.y;
-          tile.lastMoves.length = 0;
-          recordMove(global.x, global.y);
-
-          // If tile was in a slot, release it
-          if (tile.slot) {
-            tile.slot.occupiedBy = null;
-            tile.slot = null;
-          }
-        });
-
-        c.on("pointerupoutside", onUp);
-        c.on("pointerup", onUp);
-
-        // --- dragging motion
-        c.on("globalpointermove", (e: any) => {
-          if (!dragging) return;
-          const { x, y } = e.global;
-          const nx = x - dragOffset.x;
-          const ny = y - dragOffset.y;
-          c.position.set(nx, ny);
-          recordMove(x, y);
-        });
-
-        // --- record last moves for velocity detection
-        function recordMove(x: number, y: number) {
-          const now = performance.now();
-          tile.lastMoves.push({ x, y, t: now });
-          while (tile.lastMoves.length && now - tile.lastMoves[0].t > 120) {
-            tile.lastMoves.shift();
-          }
-        }
-
-        // --- calculate velocity for flick
-        function computeVelocity() {
-          if (tile.lastMoves.length < 2) return { vx: 0, vy: 0, v: 0 };
-          const a = tile.lastMoves[0];
-          const b = tile.lastMoves[tile.lastMoves.length - 1];
-          const dt = (b.t - a.t) / 1000;
-          if (dt <= 0) return { vx: 0, vy: 0, v: 0 };
-          const vx = (b.x - a.x) / dt;
-          const vy = (b.y - a.y) / dt;
-          return { vx, vy, v: Math.hypot(vx, vy) };
-        }
-
-        // --- release
-        function onUp() {
-          if (!dragging) return;
-          dragging = false;
-          c.cursor = "grab";
-          c.zIndex = 10;
-
-          const { vx, vy, v } = computeVelocity();
-          const flickThreshold = 1200; // adjust sensitivity here
-
-          if (v > flickThreshold) {
-            // flick → nearest slot or fallback
-            const targetSlot =
-              nearestEmptySlot(c.x + vx * 0.1, c.y + vy * 0.1) ||
-              firstAvailableSlot();
-            if (targetSlot) placeTileInSlot(tile, targetSlot);
-            else returnTileHome(tile);
-            return;
-          }
-
-          // normal drop → snap to hovered or nearest slot
-          const hoverSlot = hitTestEmptySlot(c.x, c.y);
-          if (hoverSlot) {
-            placeTileInSlot(tile, hoverSlot);
-            return;
-          }
-
-          const nextSlotInOrder = firstAvailableSlot();
-          if (nextSlotInOrder) placeTileInSlot(tile, nextSlotInOrder);
-          else returnTileHome(tile);
-        }
-
-        // --- unified click logic (single + double)
-        c.on("pointertap", () => {
-          const now = performance.now();
-          const diff = now - lastClickTime;
-
-          if (diff < 300) {
-            // DOUBLE CLICK detected
-            clearTimeout(clickTimeout);
-
-            // Return to rack
-            if (tile.slot) {
-              tile.slot.occupiedBy = null;
-              tile.slot = null;
-            }
-            returnTileHome(tile);
-          } else {
-            // SINGLE CLICK (delay slightly to check for double)
-            clearTimeout(clickTimeout);
-            clickTimeout = setTimeout(() => {
-              if (dragging) return;
-              if (tile.slot) return;
-              const slot = firstAvailableSlot();
-              if (slot) placeTileInSlot(tile, slot);
-            }, 300);
-          }
-
-          lastClickTime = now;
-        });
-      }
-
-
-      // =========================
-      // SLOT HELPERS
-      // =========================
-      function hitTestEmptySlot(x: number, y: number) {
-        return (
-          slots.find(
-            (s) =>
-              !s.occupiedBy &&
-              x >= s.x - 8 &&
-              x <= s.x + SLOT_SIZE + 8 &&
-              y >= s.y - 8 &&
-              y <= s.y + SLOT_SIZE + 8
-          ) || null
-        );
-      }
-
-      function nearestEmptySlot(x: number, y: number) {
-        let best: any = null,
-          bestD = Infinity;
-        for (const s of slots) {
-          if (s.occupiedBy) continue;
-          const cx = s.x + SLOT_SIZE / 2;
-          const cy = s.y + SLOT_SIZE / 2;
-          const d = (cx - x) * (cx - x) + (cy - y) * (cy - y);
-          if (d < bestD) {
-            bestD = d;
-            best = s;
-          }
-        }
-        return best;
-      }
-
-      function firstAvailableSlot() {
-        return slots.find((s) => !s.occupiedBy) || null;
-      }
-
-      // =========================
-      // TILE ACTIONS
-      // =========================
-      function placeTileInSlot(tile: any, slot: any) {
-        if (slot.occupiedBy) return;
-        slot.occupiedBy = tile;
-        tile.slot = slot;
-        tweenTo(tile.ctr, { x: slot.x, y: slot.y }, 220);
-      }
-
-      function returnTileHome(tile: any) {
-        tweenTo(tile.ctr, tile.home, 260);
-      }
 
       // =========================
       // RELAYOUT ON RESIZE
